test(storybook): split upstream account pages
